feat(audit): [AL-4353] - Atualizando rotinas de acesso dados http
